//单位id获取营业厅部分修改

+++ src/main/java/com/hbhb/cw/systemcenter/service/HallService.java
@@ -44,7 +44,7 @@ public interface HallService {
     /**
      * 通过userId去获取到这个人的营业厅列表
      */
-    Map<Long,String> selectHallByUnitId(Integer unitId);
+    Map<Integer,String> selectHallByUnitId(Integer unitId);
 
 
 }
